+++ Source/Objects/Data Processing/DataExplorer/ORDataExplorerController.h
@@ -43,7 +43,7 @@
         IBOutlet NSTextField*		multiCatalogWarningField;
 		
         BOOL                        scheduledToUpdate;
-        long                        currentSearchIndex;
+        int32_t                        currentSearchIndex;
         BOOL                        stopScan;
         BOOL                        scanInProgress;
 
@@ -92,17 +92,17 @@
 - (void) dataChanged:(NSNotification*)note;
 - (void) updateButtons;
 - (void) doUpdate;
-- (void) process:(unsigned long)row;
+- (void) process:(uint32_t)row;
 
 
 #pragma mark ¥¥¥Data Source Methods
 - (void) tableViewSelectionDidChange:(NSNotification *)aNotification;
-- (id) tableView:(NSTableView *) aTableView objectValueForTableColumn:(NSTableColumn *) aTableColumn row:(int) rowIndex;
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView;
+- (id) tableView:(NSTableView *) aTableView objectValueForTableColumn:(NSTableColumn *) aTableColumn row:(NSInteger) rowIndex;
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView;
 - (id)outlineView:(NSOutlineView *)outlineView objectValueForTableColumn:(NSTableColumn *)tableColumn byItem:(id)item; 
 - (id)outlineView:(NSOutlineView *)outlineView child:(int)index ofItem:(id)item;
 - (BOOL)outlineView:(NSOutlineView *)outlineView isItemExpandable:(id)item; 
-- (int)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item;
+- (NSUInteger)outlineView:(NSOutlineView *)outlineView numberOfChildrenOfItem:(id)item;
 
 #pragma mark ¥¥¥Delegate Methods
 - (BOOL)outlineView:(NSOutlineView *)outlineView shouldEditTableColumn:(NSTableColumn *)tableColumn item:(id)item;
