+++ Source/Objects/Hardware/Hardware Base Classes/ORCrate.h
@@ -42,7 +42,7 @@
 - (void) setAdapter:(id)anAdapter;
 - (id) controllerCard;
 - (void) doNoPowerAlert:(NSException*)exception action:(NSString*)message;
-- (unsigned long) requestGTID;
+- (uint32_t) requestGTID;
 - (NSString*) adapterArchiveKey;
 - (void) positionConnector:(ORConnector*)aConnector forCard:(id)aCard;
 - (BOOL) showLabels;
@@ -61,7 +61,7 @@
 - (void) viewChanged:(NSNotification*)aNotification;
 - (void) childChanged:(NSNotification*)aNotification;
 
-- (int)tag;
+- (NSUInteger)tag;
 - (int) crateNumber;
 - (void) setCrateNumber: (unsigned int) aCrateNumber;
 - (void) sortCards;
@@ -91,7 +91,7 @@
 @end
 
 @interface NSObject (ORCrateModel)
-- (unsigned long) requestRemoteGTID;
+- (uint32_t) requestRemoteGTID;
 @end
 
 extern NSString* ORCrateAdapterChangedNotification;
