+++ Source/Objects/Hardware/Internet/ZupSupply/ORZupController.m
@@ -49,8 +49,8 @@
     rampOpsSize		= NSMakeSize(570,750);
     blankView		= [[NSView alloc] init];
 	
-    NSString* key = [NSString stringWithFormat: @"orca.ORZup%lu.selectedtab",[model uniqueIdNumber]];
-    int index = [[NSUserDefaults standardUserDefaults] integerForKey: key];
+    NSString* key = [NSString stringWithFormat: @"orca.ORZup%u.selectedtab",[model uniqueIdNumber]];
+    NSInteger index = [[NSUserDefaults standardUserDefaults] integerForKey: key];
     if((index<0) || (index>[tabView numberOfTabViewItems]))index = 0;
     [tabView selectTabViewItemAtIndex: index];
 	
@@ -296,8 +296,8 @@
     }
     [[self window] setContentView:totalView];
             
-    NSString* key = [NSString stringWithFormat: @"orca.ORZup%lu.selectedtab",[model uniqueIdNumber]];
-    int index = [tabView indexOfTabViewItem:tabViewItem];
+    NSString* key = [NSString stringWithFormat: @"orca.ORZup%u.selectedtab",[model uniqueIdNumber]];
+    NSInteger index = [tabView indexOfTabViewItem:tabViewItem];
     [[NSUserDefaults standardUserDefaults] setInteger:index forKey:key];
     
 }
