+++ Source/Objects/Hardware/Vme/Bit3 Adaptor/ORBit3Model.h
@@ -31,39 +31,39 @@
 - (void) resetContrl;
 - (void) checkStatusErrors;
 
--(void) readLongBlock:(unsigned long *) readAddress
-									atAddress:(unsigned int) vmeAddress
-									numToRead:(unsigned int) numberLongs
+-(void) readLongBlock:(uint32_t *) readAddress
+									atAddress:(uint32_t) vmeAddress
+									numToRead:(uint32_t) numberLongs
 								   withAddMod:(unsigned short) anAddressModifier
 					   usingAddSpace:(unsigned short) anAddressSpace;
 
--(void) writeLongBlock:(unsigned long *) writeAddress
-										atAddress:(unsigned int) vmeAddress
-										numToWrite:(unsigned int) numberLongs
+-(void) writeLongBlock:(uint32_t *) writeAddress
+										atAddress:(uint32_t) vmeAddress
+										numToWrite:(uint32_t) numberLongs
 									withAddMod:(unsigned short) anAddressModifier
 						   usingAddSpace:(unsigned short) anAddressSpace;
 
 -(void) readByteBlock:(unsigned char *) readAddress
-									atAddress:(unsigned int) vmeAddress
-									numToRead:(unsigned int) numberBytes
+									atAddress:(uint32_t) vmeAddress
+									numToRead:(uint32_t) numberBytes
 								   withAddMod:(unsigned short) anAddressModifier
 					   usingAddSpace:(unsigned short) anAddressSpace;
 
 -(void) writeByteBlock:(unsigned char *) writeAddress
-										atAddress:(unsigned int) vmeAddress
-										numToWrite:(unsigned int) numberBytes
+										atAddress:(uint32_t) vmeAddress
+										numToWrite:(uint32_t) numberBytes
 									withAddMod:(unsigned short) anAddressModifier
 						   usingAddSpace:(unsigned short) anAddressSpace;
 
 -(void) readWordBlock:(unsigned short *) readAddress
-									atAddress:(unsigned int) vmeAddress
-									numToRead:(unsigned int) numberWords
+									atAddress:(uint32_t) vmeAddress
+									numToRead:(uint32_t) numberWords
 								   withAddMod:(unsigned short) anAddressModifier
 					   usingAddSpace:(unsigned short) anAddressSpace;
 
 -(void) writeWordBlock:(unsigned short *) writeAddress
-										atAddress:(unsigned int) vmeAddress
-										numToWrite:(unsigned int) numberWords
+										atAddress:(uint32_t) vmeAddress
+										numToWrite:(uint32_t) numberWords
 									withAddMod:(unsigned short) anAddressModifier
 						   usingAddSpace:(unsigned short) anAddressSpace;
 @end
